[fix] Trade 도메인 메서드 null 방어 및 타입 캐스팅 명시
Related #49

+++ src/main/java/com/ureca/snac/trade/service/TradeInitiationServiceImpl.java
@@ -86,7 +86,7 @@ public class TradeInitiationServiceImpl implements TradeInitiationService {
         CompositeBalanceResult escrowResult =
                 walletService.moveCompositeToEscrow(member.getId(), moneyToUse, pointToUse);
 
-        trade.markPaymentConfirmed(createRealTimeTradePaymentRequest.getPoint()); // Accepted -> Confirmed
+        trade.markPaymentConfirmed((int) createRealTimeTradePaymentRequest.getPoint()); // Accepted -> Confirmed
 
         // 2. 기록
         if (moneyToUse > 0) {
